Basically working now.  Login page needs cleaning up.  Default username and password:
admin
ChangeThisPassword!

To use the software as is:

ng build
node service.js

The build will build the `dist` path.  Service will serve the pages and acts as the API.

This assumes that your Silk repository is on the system that this is running on.  If it isn't, the express API has to run there and you
can serve the pages from anywhere as long as you adjust the hard coded URLs in the services (login/user and silkQuery).
x